integrate discover
